refactor: retire GpuBrain and GpuPhysics in favor of GpuKernel
Delete gpu_brain.rs (~1600 lines), gpu_physics.rs (~1100 lines), and
15 standalone WGSL shaders. GpuKernel is now the single GPU abstraction
for production, headless, and test paths.

- Add GpuKernel::is_available() (moved from GpuBrain)
- Remove gpu_brain/gpu_physics module declarations from lib.rs
- Update integration tests to use GpuKernel::is_available()
- Update agent doc comments to reference GpuKernel
- Update CLAUDE.md architecture section
